Massive refactor, super-sections now take all of the functionality
Creating a section now just requires defining the model and the row class if applicable for data sections
