+++ src/v2_samplesheet_maker/section_classes/bcl_convert_sections.py
@@ -1,0 +1,44 @@
+#!/usr/bin/env python3
+
+"""
+Define each of the available section_classes
+"""
+
+# Relative modules
+from classes.super_sections import KVSection, DataFrameSection, DataFrameSectionRow
+
+# Relative subpackages
+from ..models.bcl_convert_sections import (
+    BCLConvertDataRowModel,
+    BCLConvertSettingsSectionModel,
+    BCLConvertDataSectionModel
+)
+
+
+class BCLConvertSettingsSection(KVSection):
+    """
+    BCLConvert Settings Section
+    https://support-docs.illumina.com/SW/dragen_v42/Content/SW/DRAGEN/SampleSheet.htm
+    """
+
+    # Set model
+    _model = BCLConvertSettingsSectionModel
+
+
+class BCLConvertDataRow(DataFrameSectionRow):
+    """
+    A BCLConvert DataRow
+    """
+
+    # Set model
+    _model = BCLConvertDataRowModel
+
+
+class BCLConvertDataSection(DataFrameSection):
+    """
+    For the BCLConvert Data
+    """
+
+    # Set model
+    _model = BCLConvertDataSectionModel
+    _row_obj = BCLConvertDataRow
